Harden final test helpers
