+++ src/test_utils.rs
@@ -1,0 +1,33 @@
+use std::path::{Path, PathBuf};
+use std::sync::atomic::{AtomicU64, Ordering};
+
+static TEMP_FILE_COUNTER: AtomicU64 = AtomicU64::new(0);
+
+pub(crate) struct TempFile {
+    path: PathBuf,
+}
+
+impl TempFile {
+    pub(crate) fn new(prefix: &str, extension: &str, content: &str) -> Self {
+        let id = TEMP_FILE_COUNTER.fetch_add(1, Ordering::Relaxed);
+        let path = std::env::temp_dir().join(format!(
+            "{}-{}-{}.{}",
+            prefix,
+            std::process::id(),
+            id,
+            extension
+        ));
+        std::fs::write(&path, content).unwrap();
+        Self { path }
+    }
+
+    pub(crate) fn path(&self) -> &Path {
+        &self.path
+    }
+}
+
+impl Drop for TempFile {
+    fn drop(&mut self) {
+        let _ = std::fs::remove_file(&self.path);
+    }
+}
